Create le classi singole che implementano le operazioni grafiche.

+++ src/main/java/ch/imagik/model/ResizeInfo.java
@@ -1,24 +1,24 @@
 package ch.imagik.model;
 
 public class ResizeInfo {
-    private final double width;
-    private final double height;
+    private final int width;
+    private final int height;
     private final Integer percentage;
     public ResizeInfo(Integer percentage) {
         this.width = 0;
         this.height = 0;
         this.percentage = percentage;
     }
-    public ResizeInfo(double width, double height) {
+    public ResizeInfo(int width, int height) {
         this.width = width;
         this.height = height;
         this.percentage = 0;
     }
-    public double getHeight() {
+    public int getHeight() {
         return height;
     }
 
-    public double getWidth() {
+    public int getWidth() {
         return width;
     }
